# swagger/swagger-output.yaml
openapi: '3.0.3'

info:
  title: Board API
  description: RESTful API Board Server
  version: 1.0.0

basePath: "/"

schemes:
  - http

paths:
  /users/join:
    post:
      tags:
        - Users
      description: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 김아무개
                username:
                  type: string
                  example: kimamugae123
                password:
                  type: string
                  example: amugae123
                phoneNumber:
                  type: string
                  example: 010-8282-1234
                gender:
                  type: string
                  example: male
                birthday:
                  type: string
                  example: 2022-09-01
              required:
                - name
                - username
                - password
                - phoneNumber
                - gender
                - birthday
      responses:
        "201":
          description: "회원가입이 완료되었습니다."
        "400":
          description: "username or phoneNumber already exists"
        "500":
          description: "Server Error"
  /users/login:
    post:
      tags:
        - Users
      description: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: kimamugae123
                password:
                  type: string
                  example: amugae123
              required:
                - string
                - string
      responses:
        "201":
          description: "로그인이 완료되었습니다."
        "400":
          description: "cant find user"
        "500":
          description: "Server Error"

  /users/delete:
    delete:
      security:
        - bearerAuth: []
      tags:
        - Users
      description: 회원탈퇴
      parameters: []
      responses:
        "200":
          description: "회원탈퇴가 완료되었습니다."
        "400":
          description: "${username} doesn't exist in user table"
        "500":
          description: "Server Error"
  /users/role:
    patch:
      security:
        - bearerAuth: []
      tags:
        - Users
      description: 회원권한 변경
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: kimamugae123
                role:
                  type: string
                  example: admin
      responses:
        "200":
          description: "${username}의 회원 등급 변경이 완료되었습니다."
        "500":
          description: "Server Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT